@@ -1,25 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:planner/manager.dart';
 
-class Lines {
+import 'line.dart';
+import 'manager.dart';
+
+class Grid {
   final Manager manager;
 
-  // paint for main horizontal lines
-  Paint hpaint;
+  late Paint hPaint, vPaint;
+  final List<Line> hLines = [];
+  final List<Line> vLines = [];
 
-  // paint for all vertical lines
-  Paint vpaint;
-
-  List<Line> hlines = List<Line>();
-  List<Line> vlines = List<Line>();
-
-  Lines({@required this.manager}) {
+  Grid({required this.manager}) {
     // vertical lines: drawn after each day
     Offset vstart = Offset(manager.config.blockWidth.toDouble(), 0);
     Offset vend = Offset(
-        manager.config.blockWidth.toDouble(), manager.config.blockHeight.toDouble() * 24);
+        manager.config.blockWidth.toDouble(),
+        manager.config.blockHeight.toDouble() *
+            (manager.config.maxHour - manager.config.minHour));
     for (int i = 0; i < manager.config.labels.length; i++) {
-      vlines.add(Line(start: vstart, end: vend, manager: manager));
+      vLines
+          .add(Line(start: vstart, end: vend, controller: manager.controller));
       vstart = vstart.translate(manager.config.blockWidth.toDouble(), 0);
       vend = vend.translate(manager.config.blockWidth.toDouble(), 0);
     }
@@ -30,80 +30,69 @@ class Lines {
 
     // this is the position for the first line
     Offset hstart = Offset(0, step);
-    Offset hend =
-        Offset((manager.config.blockWidth.toDouble() * (manager.config.labels.length)), step);
+    Offset hend = Offset(
+        (manager.config.blockWidth.toDouble() * (manager.config.labels.length)),
+        step);
     int lines = (manager.config.maxHour - manager.config.minHour) * 4;
 
     for (int i = 0; i < lines; i++) {
-      hlines.add(Line(start: hstart, end: hend, manager: manager));
+      hLines
+          .add(Line(start: hstart, end: hend, controller: manager.controller));
       // increase for next 15 minute line
       hstart = hstart.translate(0, step);
       hend = hend.translate(0, step);
     }
 
-    hpaint = Paint()
-    ..color = manager.config.horizontalLineColor
-    ..style = PaintingStyle.stroke
-    ..strokeWidth = 1;
+    hPaint = Paint()
+      ..color = manager.config.horizontalLineColor
+      ..style = PaintingStyle.stroke
+      ..strokeWidth = 1;
 
-    vpaint  = Paint()
-    ..color = manager.config.verticalLineColor
-    ..style = PaintingStyle.stroke
-    ..strokeWidth = 1;
+    vPaint = Paint()
+      ..color = manager.config.verticalLineColor
+      ..style = PaintingStyle.stroke
+      ..strokeWidth = 1;
   }
 
   void draw(Canvas canvas) {
     // vertical lines can be drawn instantly
-    vlines.forEach((line) {
-      line.draw(canvas, vpaint);
-    });
+    for (var line in vLines) {
+      line.draw(canvas, vPaint);
+    }
 
     // now we need to determine the visibility for 30 minite lines, depending on zoom factor
-    double color2 = manager.zoom - 1;
+    double color2 = manager.controller.zoom - 1;
     if (color2 < 0) color2 = 0;
     if (color2 > 1) color2 = 1;
 
     // .. and set the color for this line
     Paint div2paint = Paint()
-      ..color = Color.fromARGB((color2 * 75).toInt(), 255, 255, 255)
+      ..color = Color.fromARGB((color2 * 60).toInt(), 255, 255, 255)
       ..style = PaintingStyle.stroke
       ..strokeWidth = 1;
 
     // same for the next zoomlevel, which shows 15 and 45 minute lines
-    double color3 = manager.zoom - 2;
+    double color3 = manager.controller.zoom - 2;
     if (color3 < 0) color3 = 0;
     if (color3 > 1) color3 = 1;
 
     Paint div3paint = Paint()
-      ..color = Color.fromARGB((color3 * 50).toInt(), 255, 255, 255)
+      ..color = Color.fromARGB((color3 * 30).toInt(), 255, 255, 255)
       ..style = PaintingStyle.stroke
       ..strokeWidth = 1;
 
-    // draw line
-    for (int i = 0; i < hlines.length; i++) {
+    // draw lines
+    for (int i = 0; i < hLines.length; i++) {
       if ((i + 1) % 4 == 0) {
         // hour
-        hlines[i].draw(canvas, hpaint);
+        hLines[i].draw(canvas, vPaint);
       } else if ((i + 1) % 2 == 0 && color2 != 0) {
         // half an hour
-        hlines[i].draw(canvas, div2paint);
+        hLines[i].draw(canvas, div2paint);
       } else if (color3 != 0) {
         // 15 and 45 minutes
-        hlines[i].draw(canvas, div3paint);
+        hLines[i].draw(canvas, div3paint);
       }
     }
-  }
-}
-
-class Line {
-  Offset start;
-  Offset end;
-  Manager manager;
-
-  Line({this.start, this.end, @required this.manager});
-
-  void draw(Canvas canvas, Paint paint) {
-    canvas.drawLine(manager.getScreenPosition(start),
-        manager.getScreenPosition(end), paint);
   }
 }
